moved callbacks to slack package
